[blocks] decouple blockTable from blocks view
